+++ src/lib/serialization/deserializer.ts
@@ -1,10 +1,10 @@
-import { ok } from "assert";
 import type { Application } from "../application.js";
 import {
     ArrayType,
     ConditionalType,
     DeclarationReflection,
     DocumentReflection,
+    type FileRegistry,
     IndexedAccessType,
     InferredType,
     IntersectionType,
@@ -33,11 +33,10 @@ import {
     TypeParameterReflection,
     UnionType,
     UnknownType,
-} from "../models/index.js";
-import { insertPrioritySorted } from "#utils";
+} from "#models";
+import { assert, insertPrioritySorted } from "#utils";
 import type { Logger } from "../utils/loggers.js";
 import type { JSONOutput } from "./index.js";
-import type { FileRegistry } from "../models/FileRegistry.js";
 
 export interface DeserializerComponent {
     priority: number;
@@ -59,7 +58,7 @@ export class Deserializer {
     private deferred: Array<(project: ProjectReflection) => void> = [];
     private deserializers: DeserializerComponent[] = [];
     private activeReflection: Reflection[] = [];
-    constructor(readonly application: Application) {}
+    constructor(private application: Application) {}
 
     get logger(): Logger {
         return this.application.logger;
@@ -238,10 +237,9 @@ export class Deserializer {
         options: {
             projectRoot: string;
             registry: FileRegistry;
-            addProjectDocuments?: boolean;
         },
     ): ProjectReflection {
-        ok(
+        assert(
             this.deferred.length === 0,
             "Deserializer.defer was called when not deserializing",
         );
@@ -249,9 +247,6 @@ export class Deserializer {
             name || projectObj.name,
             options.registry,
         );
-        if (options.addProjectDocuments) {
-            this.application.converter.addProjectDocuments(project);
-        }
         this.project = project;
         this.projectRoot = options.projectRoot;
         this.oldIdToNewId = { [projectObj.id]: project.id };
@@ -264,12 +259,12 @@ export class Deserializer {
             def(project);
         }
 
-        ok(
+        assert(
             this.deferred.length === 0,
             "Work may not be double deferred when deserializing.",
         );
 
-        ok(
+        assert(
             this.activeReflection.length === 0,
             "Imbalanced reflection deserialization",
         );
@@ -287,7 +282,6 @@ export class Deserializer {
         options: {
             projectRoot: string;
             registry: FileRegistry;
-            addProjectDocuments?: boolean;
         },
     ): ProjectReflection {
         if (
@@ -298,14 +292,11 @@ export class Deserializer {
         }
 
         const project = new ProjectReflection(name, options.registry);
-        if (options.addProjectDocuments) {
-            this.application.converter.addProjectDocuments(project);
-        }
         this.project = project;
         this.projectRoot = options.projectRoot;
 
         for (const proj of projects) {
-            ok(
+            assert(
                 this.deferred.length === 0,
                 "Deserializer.defer was called when not deserializing",
             );
@@ -326,12 +317,12 @@ export class Deserializer {
             for (const def of deferred) {
                 def(project);
             }
-            ok(
+            assert(
                 this.deferred.length === 0,
                 "Work may not be double deferred when deserializing.",
             );
 
-            ok(
+            assert(
                 this.activeReflection.length === 0,
                 "Imbalanced reflection deserialization",
             );
@@ -395,7 +386,7 @@ export class Deserializer {
     constructReflection<T extends JSONOutput.SomeReflection>(
         obj: T,
     ): ReflectionVariant[T["variant"]] {
-        ok(this.activeReflection.length > 0);
+        assert(this.activeReflection.length > 0);
         const result = this.reflectionBuilders[obj.variant](
             this.activeReflection[this.activeReflection.length - 1] as never,
             obj as never,