+++ src/lib/utils-common/general.ts
@@ -43,6 +43,12 @@ export function assertNever(x: never): never {
     );
 }
 
+export function assert(x: unknown, message = "Assertion failed"): asserts x {
+    if (!x) {
+        throw new Error(message);
+    }
+}
+
 export function NonEnumerable(
     _cls: unknown,
     context: ClassFieldDecoratorContext,