+++ src/lib/utils/options/tsdoc-defaults.ts
@@ -87,4 +87,5 @@ export const modifierTags = [
     "@showCategories",
     "@showGroups",
     "@useDeclaredType",
+    "@primaryExport",
 ] as const;